refactor: 예외 클래스 패키지 구조 재정비 및 출력 로직 변경
