+++ src/main/java/janggi/domain/position/Row.java
@@ -1,6 +1,6 @@
 package janggi.domain.position;
 
-import janggi.exception.RowOutOfRangeException;
+import janggi.exception.business.RowOutOfRangeException;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
