+++ src/main/java/janggi/exception/business/BusinessException.java
@@ -1,7 +1,7 @@
-package janggi.exception;
+package janggi.exception.business;
 
 public class BusinessException extends RuntimeException {
     public BusinessException(String message) {
-        super("[ERROR] " + message);
+        super(message);
     }
 }
