+++ src/main/java/janggi/exception/business/EmptyPositionException.java
@@ -1,4 +1,4 @@
-package janggi.exception;
+package janggi.exception.business;
 
 public class EmptyPositionException extends BusinessException {
     public EmptyPositionException() {
