+++ src/main/java/janggi/exception/business/InvalidMoveException.java
@@ -1,4 +1,4 @@
-package janggi.exception;
+package janggi.exception.business;
 
 public class InvalidMoveException extends BusinessException {
     public InvalidMoveException() {
