+++ src/main/java/janggi/exception/input/EmptyCoordinateException.java
@@ -1,6 +1,6 @@
-package janggi.exception;
+package janggi.exception.input;
 
-public class EmptyCoordinateException extends BusinessException {
+public class EmptyCoordinateException extends InputException {
     public EmptyCoordinateException() {
         super("좌표 값 중 하나가 비어 있습니다.");
     }
