+++ src/main/java/janggi/exception/input/EmptyInputException.java
@@ -1,6 +1,6 @@
-package janggi.exception;
+package janggi.exception.input;
 
-public class EmptyInputException extends BusinessException {
+public class EmptyInputException extends InputException {
     public EmptyInputException() {
         super("입력값이 비어있습니다.");
     }
