+++ src/main/java/janggi/exception/input/InputException.java
@@ -1,0 +1,7 @@
+package janggi.exception.input;
+
+public class InputException extends RuntimeException {
+    public InputException(String message) {
+        super(message);
+    }
+}
