+++ src/main/java/janggi/exception/input/NonNumericInputException.java
@@ -1,6 +1,6 @@
-package janggi.exception;
+package janggi.exception.input;
 
-public class NonNumericInputException extends BusinessException {
+public class NonNumericInputException extends InputException {
     public NonNumericInputException() {
         super("좌표는 숫자만 입력할 수 있습니다.");
     }
