+++ src/test/java/janggi/domain/BoardTest.java
@@ -4,7 +4,7 @@ import janggi.domain.movestrategy.JolMoveStrategy;
 import janggi.domain.position.Column;
 import janggi.domain.position.Position;
 import janggi.domain.position.Row;
-import janggi.exception.EmptyPositionException;
+import janggi.exception.business.EmptyPositionException;
 import org.junit.jupiter.api.Test;
 
 import java.util.Map;
